pyproject.toml
[project]
name = "shipment"
version = "0.1.0"
description = "shipment using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[google-genai,tools]==1.9.1",
    "pillow>=12.1.0",
    "python-barcode>=0.16.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.54.0",
]

[project.scripts]
shipment = "shipment.main:run"
run_crew = "shipment.main:run"
train = "shipment.main:train"
replay = "shipment.main:replay"
test = "shipment.main:test"
run_with_trigger = "shipment.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
